wrote FixedArray - a simple wrapper around C++ arrays; continue to implement ModuleBuilder for the factorial test

+++ src/vm/utils.h
@@ -12,8 +12,8 @@ namespace Ant {
     extern const uint64_t MB_MASKS[9];
 
 #define MB_UINT_MAX(size) MB_MASKS[(size) - 1]
-#define MB_INT_MAX(size) (MB_MASKS[(size) - 1] >> 1)
-#define MB_INT_MIN(size) (-(MB_MASKS[(size) - 1] >> 1) - 1)
+#define MB_INT_MAX(size) int64_t(MB_MASKS[(size) - 1] >> 1)
+#define MB_INT_MIN(size) (-int64_t(MB_MASKS[(size) - 1] >> 1) - 1)
 
     size_t writeMBUInt(uint64_t value, std::ostream &out);
     size_t readMBUInt(std::istream &in, uint64_t &value);
